split code into module and tool
