+++ icdiff/__init__.py
@@ -1,0 +1,9 @@
+__all__ = [
+    'diff',
+    'consolediff',
+]
+
+from icdiff.diff import diff
+from icdiff.consolediff import ConsoleDiff
+
+__version__ = "1.8.1"
